Refactor. create getTodaysDate util
