+++ src/utils/getTodaysDate.js
@@ -1,0 +1,10 @@
+function getTodaysDate() {
+  const today = new Date();
+  const year = today.getFullYear();
+  const month = String(today.getMonth() + 1).padStart(2, "0");
+  const day = String(today.getDate()).padStart(2, "0");
+
+  return `${year}/${month}/${day}`;
+}
+
+export default getTodaysDate;
